I think I missed a property in the plugin.xml ... will fix it during next merge.

# Conflicts:
#	plugins/de.cau.cs.kieler.klay.layered/plugin.xml
#	plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/GraphConfigurator.java
#	plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/intermediate/IntermediateProcessorStrategy.java
#	plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/properties/Properties.java

+++ plugins-dev/de.cau.cs.kieler.klay.codaflow/src/de/cau/cs/kieler/klay/codaflow/TIOLayoutProvider.java
@@ -4,7 +4,7 @@
  * http://www.informatik.uni-kiel.de/rtsys/kieler/
  * 
  * Copyright 2014 by
- * + Christian-Albrechts-University of Kiel
+ * + Kiel University
  *   + Department of Computer Science
  *     + Real-Time and Embedded Systems Group
  * 
@@ -65,7 +65,7 @@ public class TIOLayoutProvider extends AbstractLayoutProvider {
         // always suppress dummy port generation here
         rootLayout.setProperty(CodaflowProperties.PORT_DUMMIES, false);
 
-        moveLimit = rootLayout.getProperty(CodaflowProperties.MOVE_LIMIT);
+        moveLimit = rootLayout.getProperty(CodaflowProperties.MOVE_LIMIT).doubleValue();
 
         // calculate margins of the nodes
         KGraphAdapter adapter = KGraphAdapters.adapt(parentNode);
